taskflow-sre-sandbox/apps/worker/requirements.txt
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
prometheus_client==0.19.0
